Move PaymentSheet ui utils
