+++ paymentsheet/src/main/java/com/stripe/android/paymentelement/embedded/manage/ManageActivity.kt
@@ -27,6 +27,7 @@ import com.stripe.android.paymentelement.embedded.EmbeddedSelectionHolder
 import com.stripe.android.paymentsheet.CustomerStateHolder
 import com.stripe.android.paymentsheet.R
 import com.stripe.android.paymentsheet.ui.PaymentSheetTopBar
+import com.stripe.android.paymentsheet.utils.renderEdgeToEdge
 import com.stripe.android.ui.core.elements.H4Text
 import com.stripe.android.uicore.StripeTheme
 import com.stripe.android.uicore.elements.bottomsheet.rememberStripeBottomSheetState
@@ -64,6 +65,8 @@ internal class ManageActivity : AppCompatActivity() {
             finish()
             return
         }
+
+        this.renderEdgeToEdge()
 
         viewModel.component.inject(this)
