+++ paymentsheet/src/main/java/com/stripe/android/paymentsheet/paymentdatacollection/bacs/BacsMandateConfirmationActivity.kt
@@ -1,7 +1,6 @@
 package com.stripe.android.paymentsheet.paymentdatacollection.bacs
 
 import android.app.Activity
-import android.os.Build
 import android.os.Bundle
 import androidx.activity.addCallback
 import androidx.activity.compose.setContent
@@ -9,13 +8,13 @@ import androidx.activity.viewModels
 import androidx.appcompat.app.AppCompatActivity
 import androidx.compose.material.ExperimentalMaterialApi
 import androidx.compose.runtime.LaunchedEffect
-import androidx.core.view.WindowCompat
 import com.stripe.android.common.ui.BottomSheetScaffold
 import com.stripe.android.common.ui.ElementsBottomSheetLayout
 import com.stripe.android.paymentsheet.parseAppearance
 import com.stripe.android.paymentsheet.paymentdatacollection.bacs.BacsMandateConfirmationViewAction.OnBackPressed
 import com.stripe.android.paymentsheet.ui.PaymentSheetTopBar
 import com.stripe.android.paymentsheet.ui.PaymentSheetTopBarState
+import com.stripe.android.paymentsheet.utils.renderEdgeToEdge
 import com.stripe.android.uicore.StripeTheme
 import com.stripe.android.uicore.elements.bottomsheet.rememberStripeBottomSheetState
 import com.stripe.android.uicore.utils.fadeOut
@@ -36,7 +35,7 @@ internal class BacsMandateConfirmationActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
 
-        renderEdgeToEdge()
+        this.renderEdgeToEdge()
 
         onBackPressedDispatcher.addCallback {
             viewModel.handleViewAction(OnBackPressed)
@@ -91,13 +90,5 @@ internal class BacsMandateConfirmationActivity : AppCompatActivity() {
     override fun finish() {
         super.finish()
         fadeOut()
-    }
-
-    private fun renderEdgeToEdge() {
-        if (Build.VERSION.SDK_INT < Build.VERSION_CODES.R) {
-            return
-        }
-
-        WindowCompat.setDecorFitsSystemWindows(window, false)
     }
 }
