+++ paymentsheet/src/main/java/com/stripe/android/paymentsheet/ui/BaseSheetActivity.kt
@@ -1,11 +1,10 @@
 package com.stripe.android.paymentsheet.ui
 
-import android.os.Build
 import android.os.Bundle
 import androidx.activity.addCallback
 import androidx.appcompat.app.AppCompatActivity
-import androidx.core.view.WindowCompat
 import com.stripe.android.paymentsheet.LinkHandler
+import com.stripe.android.paymentsheet.utils.renderEdgeToEdge
 import com.stripe.android.paymentsheet.viewmodels.BaseSheetViewModel
 import com.stripe.android.uicore.utils.fadeOut
 
@@ -26,7 +25,7 @@ internal abstract class BaseSheetActivity<ResultType> : AppCompatActivity() {
             return
         }
 
-        renderEdgeToEdge()
+        this.renderEdgeToEdge()
 
         onBackPressedDispatcher.addCallback {
             viewModel.handleBackPressed()
@@ -36,13 +35,5 @@ internal abstract class BaseSheetActivity<ResultType> : AppCompatActivity() {
     override fun finish() {
         super.finish()
         fadeOut()
-    }
-
-    private fun renderEdgeToEdge() {
-        if (Build.VERSION.SDK_INT < 30) {
-            return
-        }
-
-        WindowCompat.setDecorFitsSystemWindows(window, false)
     }
 }
