+++ paymentsheet/src/main/java/com/stripe/android/paymentsheet/utils/EdgeToEdge.kt
@@ -1,14 +1,13 @@
 package com.stripe.android.paymentsheet.utils
 
-import androidx.compose.foundation.layout.Spacer
-import androidx.compose.foundation.layout.requiredHeight
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.res.dimensionResource
-import com.stripe.android.paymentsheet.R
+import android.app.Activity
+import android.os.Build
+import androidx.core.view.WindowCompat
 
-@Composable
-internal fun PaymentSheetContentPadding() {
-    val bottomPadding = dimensionResource(R.dimen.stripe_paymentsheet_button_container_spacing_bottom)
-    Spacer(modifier = Modifier.requiredHeight(bottomPadding))
+internal fun Activity.renderEdgeToEdge() {
+    if (Build.VERSION.SDK_INT < Build.VERSION_CODES.R) {
+        return
+    }
+
+    WindowCompat.setDecorFitsSystemWindows(window, false)
 }
